### 2025-05-23 #8265 @editorialbot
[PRE REVIEW]: fireexposuR: An R package for computing and visualizing wildfire exposure
**Submitting author:** @heyairf (<a href="http://orcid.org/0000-0002-9842-7648">Air M. Forbes</a>)
**Repository:** https://github.com/ropensci/fireexposuR
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
