### 2025-06-21 #8265 @mengqi-z
👋 @thomasarsouze @yotarazona @nickbearman - Would any of you be interested and available to review this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
