### 2025-06-24 #8479 @editorialbot
[REVIEW]: fireexposuR: An R package for computing and visualizing wildfire exposure
**Submitting author:** @heyairf (<a href="http://orcid.org/0000-0002-9842-7648">Air M. Forbes</a>)
**Repository:** https://github.com/ropensci/fireexposuR
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
