### 2025-07-16 ropensci/fireexposuR#11 @mstorey87
dependencies
Hi there, just a question in relation to openjournals/joss-reviews#8479
when I run "install.packages("fireexposuR", repos = "https://ropensci.r-universe.dev")", the package seems to install ok but i get the warning:

Warning in install.packages :
  dependencies ‘MultiscaleDTM’, ‘tidyterra’ are not available

I can install those packages independently though. Any idea what is going on?
