### 2025-07-17 ropensci/fireexposuR#11 @mstorey87
I just installed again and the dependencies were installed ok
