### 2025-07-18 ropensci/fireexposuR#11 @heyairf
Hi @mstorey87 ,
Not sure what is going on there... perhaps could due to installing through R-universe instead of CRAN? i will test both... and also update the readme to have both options now that the package is on CRAN
