### 2025-07-20 ropensci/fireexposuR#11 @mstorey87
Not sure either, but seems all good now. I'll close this.
