### 2025-07-21 #8479 @mstorey87
Hi @mengqi-z , in relation to the checklist item: "Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution."
Where should dependencies be listed? In documentation here https://github.com/ropensci/fireexposuR?
