### 2025-08-04 ropensci/fireexposuR#12 @thomasarsouze
Installation issue
As part of the review process https://github.com/openjournals/joss-reviews/issues/8479

Similar to #11, I also have a warning with dependencies issues when following installation procedure. However, this only happens the first time, hence difficult to reproduce. Indeed probably due to the `ropensci` repository.

I don't know how releases on `ropensci` work, but you mention `development version`. Is it really ? Usualy 'devlopment versions" are direct clones of git repository. Maybe now that you also have a version on `cran`, you should detail how to install such a 'stable' version, and also probably insert a badge with the `cran` version :

[![](https://www.r-pkg.org/badges/version/fireexposuR?color=green)](https://cran.r-project.org/package=fireexposuR)
