### 2025-08-05 ropensci/fireexposuR#14 @thomasarsouze
Merge `fire_exp` and `fire_exp_adjust` functions ?
As part of the review process https://github.com/openjournals/joss-reviews/issues/8479

Code in these 2 functions is too similar and almost duplicated (with caveats associated, such as check https://github.com/ropensci/fireexposuR/blob/b1863d237672e7ba1d5ce0e992fb1583eccb140d/R/fire_exp_adjust.R#L64C1-L65C48 done in `fire_exp_adjust` and not in `fire_exp`).

I would recommend to merge the two functions and deal with different `tdist` type of input.
