### 2025-08-05 ropensci/fireexposuR#15 @thomasarsouze
Redondant columns in `fire_exp_summary`
As part of the review process https://github.com/openjournals/joss-reviews/issues/8479

Both `n` and `npixels` present in the output of `fire_exp_summary` refer to the same value. Please drop the `n`  column.
