### 2025-08-05 ropensci/fireexposuR#17 @thomasarsouze
Use consistent API : unfamiliar use of `fire_exp_extract_vis`
As part of the review process https://github.com/openjournals/joss-reviews/issues/8479

Not really a big issue here, and more a recommendation than a request, but `fire_exp_extract_vis` is not really in line with other functions naming.

`Vis` stands for visualization but in other function you:
- likely refer to `summarize` for the production of a table as an output
- have a proper map function to produce a map.
