### 2025-08-05 ropensci/fireexposuR#17 @thomasarsouze
edit :

I realize that you use the same function behavior for `fire_exp_dir_multi`, so same comment applies.
