### 2025-08-05 ropensci/fireexposuR#18 @thomasarsouze
Complete the `Get started` vignette
As part of the review process https://github.com/openjournals/joss-reviews/issues/8479

The `Get started` vignette is very detailed and clearly documented, however, I think it would benefit from have outputs of functions included, in other words : run the code examples.

I think you didn't want to include a large raster for the 'local' scale, but even an example on a very short domain would help to fully apprehend the functionalities of your package.
