### 2025-08-05 ropensci/fireexposuR#19 @thomasarsouze
Update `Preparing input data` vignette
As part of the review process https://github.com/openjournals/joss-reviews/issues/8479

For the release in JOSS, it would be nice to have a final (or let's say 'more achieved') version of the `Preparing input data` vignette.

I understand that most of the relevant information is in Forbes AM, Beverly JL (2024), so I let you decide whether you want to:
- resume some of the main points of the paper here,
- have 1 or 2 most common examples on how to generate data
- simply refer to the paper

I think all options are good, but depend on how you think this will most help the user.
