### 2025-08-22 #8479 @heyairf
Hi @mengqi-z! I am making some changes to address the opened issues (and also doing some new CRAN test failure troubleshooting) locally... I am aiming to push these changes sometime next week.
