### 2025-08-22 b2dd4c1
remove redundant column
