### 2025-08-22 2c7d47d
split fire_exp_extract_vis into two different functions to keep function naming consistent
