+++ R/fire_exp_extract_map.R
@@ -1,8 +1,8 @@
-#' Visualize exposure to values in a summary table or map (deprecated)
+#' Visualize exposure to values in a map
 #'
-#' @description This function still works, but please update your code to
-#' use the new [fire_exp_extract_summary()] or [fire_exp_extract_map()] instead.
-#' This function will be removed in the future.
+#' @description `fire_exp_extract_map()` standardizes the visualization of
+#' outputs from [fire_exp_extract()] as a map by classifying
+#' exposure into predetermined exposure classes.
 #'
 #' @details
 #' This function visualizes the outputs from [fire_exp_extract()] with classes.
@@ -51,8 +51,6 @@
 #' @param method character, either `"max"` or `"mean"`. If `values_ext` are
 #' polygons the default is `"max"`.This parameter is ignored when `values_ext`
 #' are point features.
-#' @param map Boolean. When `TRUE`, a map is returned as a ggplot object. The
-#' default is `FALSE`.
 #' @param zoom_level (Optional). Numeric. Ignored when `map = FALSE`. set the
 #' zoom level for the base map tile. See details. Defaults if:
 #' * `classify = "local"` or `"custom"` the zoom level default is `12`
@@ -62,8 +60,7 @@
 #'
 #'
 #'
-#' @return a summary table is returned as a data frame object, Unless:
-#' `map = TRUE`: a ggplot object
+#' @return a summary table is returned as a ggplot object
 #'
 #' @export
 #'
@@ -87,25 +84,15 @@
 #'
 #' values_exp <- fire_exp_extract(exposure, points)
 #'
-#' # summarize example points in a table
-#' fire_exp_extract_vis(values_exp, classify = "local")
-#'
 #' # visualize example points in standardized map
-#' fire_exp_extract_vis(values_exp, map = TRUE)
+#' fire_exp_extract_map(values_exp)
 #'
-fire_exp_extract_vis <- function(values_ext,
+fire_exp_extract_map <- function(values_ext,
                                  classify = c("local", "landscape", "custom"),
                                  class_breaks,
                                  method = c("max", "mean"),
-                                 map = FALSE,
                                  zoom_level,
                                  title = "Classified Exposure to Values") {
-  if (map == TRUE) {
-    .Deprecated("fire_exp_extract_map")
-  } else {
-    .Deprecated("fire_exp_extract_summary")
-  }
-
   ext <- values_ext
   stopifnot("`values_ext` must be a SpatVector of point or polygon features"
             = (class(ext) == "SpatVector" &&
@@ -179,69 +166,60 @@ fire_exp_extract_vis <- function(values_ext,
 
   ext_class$class_range <- factor(ext_class$class_range, levels = names(lut2))
 
-  if (map == TRUE) {
 
-    n_color <- length(class_breaks)
+  n_color <- length(class_breaks)
 
-    cols <- c("grey40", tidyterra::whitebox.colors(n_color,
-                                                   palette = "bl_yl_rd"))
-    names(cols) <- class_labels
+  cols <- c("grey40", tidyterra::whitebox.colors(n_color,
+                                                 palette = "bl_yl_rd"))
+  names(cols) <- class_labels
 
-    v <- terra::project(ext_class, "EPSG: 3857") %>%
-      tidyr::drop_na("class")
-    e <- terra::rescale(v, 1.5)
-    tile <- maptiles::get_tiles(e, "Esri.WorldImagery", zoom = zoom_level) %>%
-      terra::crop(e)
+  v <- terra::project(ext_class, "EPSG: 3857") %>%
+    tidyr::drop_na("class")
+  e <- terra::rescale(v, 1.5)
+  tile <- maptiles::get_tiles(e, "Esri.WorldImagery", zoom = zoom_level) %>%
[…]
