+++ NAMESPACE
@@ -8,6 +8,8 @@ export(fire_exp_dir_map)
 export(fire_exp_dir_multi)
 export(fire_exp_dir_plot)
 export(fire_exp_extract)
+export(fire_exp_extract_map)
+export(fire_exp_extract_summary)
 export(fire_exp_extract_vis)
 export(fire_exp_map_class)
 export(fire_exp_map_cont)

[…]
