### 2025-08-22 ddcbe51
merge functionality of `fire_exp_adjust()` into `fire_exp()`
