+++ R/fire_exp_adjust.R
@@ -1,7 +1,9 @@
 #' Compute the wildfire exposure metric with custom transmission distances
+#' (Deprecated)
 #'
-#' @description For advanced users. Adjust the transmission distances from the
-#' defaults used in [fire_exp()].
+#' @description This function still works, but will be removed in future
+#' versions of the package. The same functionality is now included in
+#'  [fire_exp()]. .
 #'
 #' @details
 #' If the transmission distances from the wildfire exposure literature are not
@@ -47,6 +49,9 @@
 #' exposure800 <- fire_exp_adjust(hazard, tdist = 800)
 #'
 fire_exp_adjust <- function(hazard, tdist, no_burn) {
+
+  .Deprecated("fire_exp")
+
   stopifnot("`hazard` must be a SpatRaster object"
             = class(hazard) == "SpatRaster",
             "`hazard` layer must have values between 0-1"
