+++ R/fire_exp.R
@@ -25,7 +25,7 @@
 #' source. If the default distances do not accurately represent fire behaviour
 #' in your area of interest you can adjust them with [fire_exp_adjust()].
 #'
-#' ### Radiant heat
+#' ### Radiant heat (30 m)
 #'
 #' The default transmission distance for radiant heat is 30 meters. At this
 #' scale of wildfire transmission wildfire can spread from direct flame contact
@@ -38,7 +38,7 @@
 #' reviewed fire-spot distance observations and applied predictive models to
 #' verify these distances were a reasonable assumption.
 #'
-#' #### Short-range embers
+#' #### Short-range embers (100 m)
 #'
 #' The default transmission distance for short-range embers is 100 meters. Some
 #' fuel types have the potential to produce embers over short distances. An
@@ -47,7 +47,7 @@
 #' will also be hazardous to transmit fire via radiant heat or direct flame
 #' contact.
 #'
-#' #### Long-range embers
+#' #### Long-range embers (500 m)
 #'
 #' The default transmission distance for long-range embers is 500 meters.
 #' Some fuel types have the potential to transmit embers across large distances.
@@ -63,8 +63,19 @@
 #'
 #' The 500 meter spread distance has been further validated across Alberta
 #' (Beverly et al. 2021), in Alaska (Schmidt et al. 2024), across the entire
-#' Canadian landbase (manuscript in preparation), and in Portugal (manuscript
-#' in preparation).
+#' Canadian landbase (manuscript in preparation), and in Portugal (Khan et al.
+#' 2025).
+#'
+#'
+#' ### Setting a custom transmission distance
+#'
+#' If the transmission distances from the wildfire exposure literature are not
+#' representative of the wildland fuels in your area of interest, this function
+#' can be used to change the transmission distance to a custom distance. It is
+#' highly recommended that any exposure layers produced with this function are
+#' validated with observed fire history using the [fire_exp_validate()]
+#' function. See methods of validation in Beverly et al. 2021, Schmidt et al.
+#' 2024, and Khan et al. 2025 for more discussion on validation.
 #'
 #' ## Technical
 #'
@@ -92,10 +103,13 @@
 #'
 #' A separate hazard raster should be prepared for each of the transmission
 #' distances of interest. There are also minimum spatial resolution and extent
-#' requirements for each transmission distances.
+#' requirements for each transmission distances.For a specified transmission
+#' distance, the spatial resolution must be at least 3 times finer. For example,
+#'  for a transmission distance of 300 meters the input data should have a
+#'  resolution of 100 meters or finer.
 #'
 #' Long-range embers:
-#' -  minimum raster resolution is 150 meters
+#' -  minimum raster resolution is ~ 150 meters
 #' -  The dimensions of the data must be wider/taller than 1000 meters because
 #' 500 meters of data will be lost along the perimeter due to edge effects
 #'
@@ -112,8 +126,10 @@
 #' ### Spatial Reference
 #'
 #' The exposure raster will be returned in the same CRS as the input hazard
-#' layer. A crs must be defined if the outputs will be used in other functions
-#' in this package.
+#' layer. A crs must be defined if the output will be used as an input to other
+#' functions in this package.
+#'
+#'
 #'
 #' @references
 #' Beverly JL, McLoughlin N, Chapman E (2021) A simple metric of landscape fire
@@ -131,6 +147,11 @@
[…]
