### 2025-08-22 c61e9cf
add more details about deprecation
