### 2025-08-22 5f9eb6e
add Khan et al citations
