### 2025-08-22 46ffc0f
update prometheus issue to include burn-p3 (and include links for clarity)
