### 2025-08-22 162816c
add CRAN badge
