### 2025-08-28 2394677
add deprecated message
