### 2025-08-28 729c845
add deprecated message
