+++ R/fire_exp_dir_plot.R
@@ -36,7 +36,7 @@
 #' hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 #'
 #' # generate an example point
-#' point_wkt <- "POINT (400000 6050000)"
+#' point_wkt <- "POINT (345000 5876000)"
 #' point <- terra::vect(point_wkt, crs = hazard)
 #'
 #' # compute exposure metric
@@ -48,7 +48,7 @@
 #'
 #' # radial plot
 #' fire_exp_dir_plot(transects)
-#'
+
 #' # customize labels
 #' fire_exp_dir_plot(transects, labels = c("5 km", "10 km", "15 km"))
 #'
