+++ R/fire_exp_dir_map.R
@@ -44,7 +44,7 @@
 #' hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 #'
 #' # generate an example point
-#' point_wkt <- "POINT (400000 6050000)"
+#' point_wkt <- "POINT (345000 5876000)"
 #' point <- terra::vect(point_wkt, crs = hazard)
 #'
 #' # compute exposure metric
