+++ R/fire_exp_map_class.R
@@ -72,12 +72,9 @@
 #' hazard_file_path <- "extdata/hazard.tif"
 #' hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 #'
-#' # read example area of interest geometry
-#' geom_file_path <- "extdata/polygon_geometry.csv"
-#' geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
-#'
-#' # generate example area of interest polygon with geometry
-#' aoi <- terra::vect(as.matrix(geom), "polygons", crs = hazard)
+#' # read example area of interest
+#' polygon_path <- system.file("extdata", "polygon.shp", package ="fireexposuR")
+#' aoi <- terra::vect(polygon_path)
 #'
 #' # compute exposure
 #' exposure <- fire_exp(hazard)
