### 2025-08-28 1e4ea21
move deprecated calls to reduce test clutter
