### 2025-08-28 05838ee
new function (consolidate fire_exp_map_class and _cont)
