+++ NAMESPACE
@@ -11,6 +11,7 @@ export(fire_exp_extract)
 export(fire_exp_extract_map)
 export(fire_exp_extract_summary)
 export(fire_exp_extract_vis)
+export(fire_exp_map)
 export(fire_exp_map_class)
 export(fire_exp_map_cont)
 export(fire_exp_summary)
