### 2025-08-28 51b352c
new function (split fire_exp_ext_vis into _map and _summary)
