+++ man/fire_exp_dir_plot.Rd
@@ -39,7 +39,7 @@ hazard_file_path <- "extdata/hazard.tif"
 hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 
 # generate an example point
-point_wkt <- "POINT (400000 6050000)"
+point_wkt <- "POINT (345000 5876000)"
 point <- terra::vect(point_wkt, crs = hazard)
 
 # compute exposure metric
@@ -51,7 +51,6 @@ transects <- fire_exp_dir(exposure, point)
 
 # radial plot
 fire_exp_dir_plot(transects)
-
 # customize labels
 fire_exp_dir_plot(transects, labels = c("5 km", "10 km", "15 km"))
 
