+++ man/fire_exp_extract_vis.Rd
@@ -97,12 +97,9 @@ more information on zoom levels.
 hazard_file_path <- "extdata/hazard.tif"
 hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 
-# read example area of interest geometry
-geom_file_path <- "extdata/polygon_geometry.csv"
-geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
-
-# generate an area of interest polygon with the geometry
-aoi <- terra::vect(as.matrix(geom), "polygons", crs = hazard)
+# read example area of interest
+polygon_path <- system.file("extdata", "polygon.shp", package ="fireexposuR")
+aoi <- terra::vect(polygon_path)
 
 # generate random points within the aoi polygon
 points <- terra::spatSample(aoi, 100)
