### 2025-08-28 4aa94e1
add deprecated details to docs
