### 2025-08-28 dfb81ac
update readme with new example data and function names/params
