+++ README.Rmd
@@ -182,15 +182,15 @@ The `aoi` layer is a polygon representing a localized area of interest
 campground, etc.) shown in red.
 
 ```{r map_inputs, echo = FALSE}
-plot(as.factor(hazard))
+plot(hazard)
 plot(aoi, add = T, border = "red", lwd = 3)
 ```
 
 ### Compute exposure
 
 ```{r compute}
-# compute long-range ember exposure by setting transmission distance to "l"
-exposure <- fire_exp(hazard, tdist = "l")
+# compute long-range ember exposure
+exposure <- fire_exp(hazard, t_dist = 500)
 
 # compute directional exposure toward the value with default parameters
 dir_exposure <- fire_exp_dir(exposure, aoi)
@@ -198,8 +198,11 @@ dir_exposure <- fire_exp_dir(exposure, aoi)
 
 These objects can be exported using the terra library if the user
 prefers visualizing and conducting further analysis outside of the R
-environment (e.g. a GIS). - The `exposure` layer can be exported as a
-raster - The `dir_exposure` layer can be exported as a shapefile
+environment (e.g. a GIS).
+
+- The `exposure` layer can be exported as a
+raster 
+- The `dir_exposure` layer can be exported as a shapefile
 
 ### Visualize exposure
 
@@ -208,10 +211,10 @@ package.
 
 ```{r visualize}
 # map the full extent of the exposure raster with a continuous scale
-fire_exp_map_cont(exposure)
+fire_exp_map(exposure)
 
-# map exposure classes within the area of interest with a base map
-fire_exp_map_class(exposure, aoi, classify = "landscape", zoom_level = 13)
+# map exposure classes within the area of interest
+fire_exp_map(exposure, aoi, classify = "local")
 
 # map the directional exposure transects toward the area of interest
 fire_exp_dir_map(dir_exposure, aoi)
