+++ R/fire_exp_dir_map.R
@@ -6,28 +6,20 @@
 #' @details
 #' This function returns a standardized map with basic cartographic elements.
 #'
-#' The plot is returned as a ggplot object which can be exported/saved to
-#' multiple image file formats.
+#' The plot is returned as a tmap object which can be further customized using
+#' tmap commands or exported/saved to multiple image file formats. See
+#' [tmap::tmap_save()] for export details.
 #'
 #' ## Spatial reference
 #'
-#' This function dynamically pulls map tiles for a base map. The inputs are
-#' projected to WGS 84/Pseudo-Mercator ([EPSG:3857](https://epsg.io/3857)) to
-#' align them with the map tiles.
+#' This function dynamically pulls map tiles for a base map. The crs is set
+#' automatically. See [tmap::tm_crs()] for details.
 #'
-#' ## Zoom level
-#' The map tile zoom level may need to be adjusted. If the base map is blurry,
-#' increase the zoom level. Higher zoom levels will slow down the function, so
-#' only increase if necessary. Reference the
-#' [OpenStreetMap Wiki](https://wiki.openstreetmap.org/wiki/Zoom_levels) for
-#' more information on zoom levels.
 #'
 #'
 #' @param transects SpatVector. Output from [fire_exp_dir()]
 #' @param value (Optional) SpatVector. Adds the value to the map. Use the same
 #' value feature used to generate the transects with [fire_exp_dir()]
-#' @param zoom_level (Optional). Numeric. set the zoom level for the base map
-#' tile. See details. The default is `10`.
 #' @param labels (Optional) a vector of three strings. Custom formatting for the
 #' distances in the legend. If left blank, the function will automatically label
 #' the distances in meters.
@@ -35,7 +27,7 @@
 #' `"Directional Vulnerability"`
 #'
 #'
-#' @return a map of directional exposure transects as a ggplot object
+#' @return a map of directional exposure transects as a tmap object
 
 #' @export
 #' @examples
@@ -51,24 +43,19 @@
 #' exposure <- fire_exp(hazard)
 #'
 #' # generate transects
-#' transects <- fire_exp_dir(exposure, point)
+#' transects <- fire_exp_dir(exposure, point, interval = 5)
 #'
-#' # map with customized distance labels
-#' fire_exp_dir_map(transects, labels = c("5 km", "10 km", "15 km"),
-#'                  zoom_level = 9)
+#' fire_exp_dir_map(transects)
 #'
 
 
 
 fire_exp_dir_map <- function(transects,
                              value,
-                             zoom_level = 10,
                              labels,
                              title = "Directional Vulnerability") {
   stopifnot("`transects` must be a SpatVector object"
             = class(transects) == "SpatVector",
-            "`zoom_level` must be a number"
-            = class(zoom_level) == "numeric",
             "`title` must be a character string"
             = class(title) == "character")
 
@@ -87,68 +74,37 @@ fire_exp_dir_map <- function(transects,
             paste(labels[1], "to", labels[2]),
             paste(labels[2], "to", labels[3]))
 
-  t <- tidyterra::filter(transects, .data$viable == 1) %>%
-    terra::project("EPSG:3857")
+  cols <- c("darkred", "darkorange", "yellow2")
 
-  e <- transects %>%
-    terra::project("EPSG:3857") %>%
-    terra::rescale(1.3)
+  t <- tidyterra::filter(transects, .data$viable == 1)
 
[…]
