+++ R/fire_exp_validate.R
@@ -62,9 +62,8 @@
 #' hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 #'
 #' # generate example non-burnable cells data
-#' geom_file_path <- "extdata/polygon_geometry.csv"
-#' geom <- read.csv(system.file(geom_file_path, package = "fireexposuR"))
-#' polygon <- terra::vect(as.matrix(geom), "polygons", crs = hazard)
+#' polygon_path <- system.file("extdata", "polygon.shp", package ="fireexposuR")
+#' polygon <- terra::vect(polygon_path)
 #' no_burn <- terra::rasterize(polygon, hazard)
 #'
 #' # generate example fire polygons by buffering random points
