+++ vignettes/fireexposuR.Rmd
@@ -68,13 +68,11 @@ The simplest use case of this package is computing the exposure metric. All of t
 
 Analysis functions: 
 
--  `fire_exp()` : Compute exposure with default transmission distances
--  `fire_exp_adjust()` : Compute exposure with custom transmission distances
+-  `fire_exp()` : Compute exposure for a given transmission distance
 
 Visualization functions:
 
--  `fire_exp_map_cont()` : Visualize exposure mapped with a continuous scale
--  `fire_exp_map_class()` : Visualize exposure mapped with a classified scale
+-  `fire_exp_map()` : Visualize exposure mapped with a continuous scale or classified scale
 -  `fire_exp_summary()` : Visualize exposure classes in a summary table
 
 Required input data:
@@ -110,10 +108,10 @@ hazard_long <- rast("long_range_hazard.tif")
 hazard_short <- rast("short_range_hazard.tif")
 
 # compute long-range ember exposure using the long-range hazard layer
-exposure_long <- fire_exp(hazard_long, tdist = "l")
+exposure_long <- fire_exp(hazard_long, t_dist = 500)
 
 # compute short-range ember exposure using the short-range hazard layer
-exposure_short <- fire_exp(hazard_short, tdist = "s")
+exposure_short <- fire_exp(hazard_short, t_dist = 100)
 
 # Export the results as a .tif file using terra
 writeRaster(exposure_long, "exposure_long.tif")
@@ -193,7 +191,8 @@ Analysis functions:
 
 Visualization functions:
 
--  `fire_exp_map_extract_vis()` : Visualize classified exposure to values in a map or summary table
+-  `fire_exp_map_extract_map()` : Visualize classified exposure to values in a map
+-  `fire_exp_map_extract_summary()` : Visualize classified exposure to values in a summary table
 -  `fire_exp_map_class()` : Visualize exposure mapped with a classified scale 
 -  `fire_exp_map_summary()` : Visualize exposure classes in a summary table 
 
@@ -229,10 +228,10 @@ hazard_long <- rast("long_range_hazard.tif")
 hazard_short <- rast("short_range_hazard.tif")
 
 # compute long-range ember exposure using the long-range hazard layer
-exposure_long <- fire_exp(hazard_long, tdist = "l")
+exposure_long <- fire_exp(hazard_long, t_dist = 500)
 
 # compute short-range ember exposure using the short-range hazard layer
-exposure_short <- fire_exp(hazard_short, tdist = "s")
+exposure_short <- fire_exp(hazard_short, t_dist = 100)
 
 # read polygon feature vector of an area of interest boundary
 # (e.g., a built environment)
@@ -252,19 +251,19 @@ values_exposure_short <- fire_exp_extract(exposure_short, multiple_values)
 fire_exp_summary(exposure_long, area_of_interest, classify = "local")
 
 # Visualize long-range ember exposure classes in the aoi with a map
-fire_exp_map_class(exposure_long, area_of_interest, classify = "local")
+fire_exp_map(exposure_long, area_of_interest, classify = "local")
 
 # Visualize long-range ember exposure classes to values in summary table
-fire_exp_extract_vis(values_exposure_long, classify = "local")
+fire_exp_extract_summary(values_exposure_long, classify = "local")
 
 # Visualize long-range ember exposure classes to values in map
-fire_exp_extract_vis(values_exposure_long, classify = "local", map = TRUE)
+fire_exp_extract_map(values_exposure_long, classify = "local")
 
 # Repeat the visualizations for short-range ember exposure
[…]
