### 2025-08-28 b40c685
aesthetics update of map components
