+++ R/fire_exp_map.R
@@ -175,12 +175,15 @@ fire_exp_map <- function(exposure, aoi, classify,
 
 
   plt <- plt + tmap::tm_basemap("Esri.WorldImagery") +
-    tmap::tm_credits("Basemap: Esri World Imagery") +
+    tmap::tm_credits("Basemap: Esri World Imagery",
+                     color = "white") +
     tmap::tm_title(title) +
     tmap::tm_compass(
       type = "arrow",
       position = c("LEFT", "BOTTOM"),
-      color.light = "white"
+      color.light = "black",
+      color.dark = "white",
+      text.color = "white"
     ) +
     tmap::tm_scalebar(position = tmap::tm_pos_out("center", "bottom"),
                       text.size = 0.9) +
