### 2025-08-28 fd02459
make maps look better
