### 2025-08-29 2f4c2a3
split fire_exp_dir_multi
