+++ NAMESPACE
@@ -6,6 +6,7 @@ export(fire_exp_adjust)
 export(fire_exp_dir)
 export(fire_exp_dir_map)
 export(fire_exp_dir_multi)
+export(fire_exp_dir_multi_plot)
 export(fire_exp_dir_plot)
 export(fire_exp_extract)
 export(fire_exp_extract_map)

[…]
