### 2025-09-03 37cc056
add clarity to error message
