### 2025-09-03 426f876
map legend adjustments
