### 2025-09-10 8e16c1f
move installation instructions higher up
