### 2025-09-10 d3e531e
add new function
