+++ _pkgdown.yml
@@ -11,6 +11,8 @@ reference:
   - fire_exp_dir
   - fire_exp_extract
   - fire_exp_validate
+  - fire_exp_dir_multi
+
 - title: "Visualize"
   desc: >
     Functions for visualizing wildfire exposure outputs in R
@@ -21,8 +23,9 @@ reference:
   - fire_exp_extract_map
   - fire_exp_dir_map
   - fire_exp_dir_plot
-  - fire_exp_dir_multi
   - fire_exp_validate_plot
+  - fire_exp_dir_multi_plot
+
 
 - title: "Deprecated"
   desc: >
