### 2025-09-10 ba8f182
round proportion in output so its easier to read
