### 2025-09-10 d12c10e
update example data and change example code so examples have outputs and interpretation
