+++ vignettes/fireexposuR.Rmd
@@ -10,7 +10,8 @@ vignette: >
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  comment = "#>"
+  comment = "#>",
+  warning = FALSE, message = FALSE
 )
 ```
 
@@ -92,31 +93,62 @@ Additional data for visualization:
 
 #### Example workflow
 
-This example code demonstrates how to use the package to compute wildfire exposure at two different scales of wildfire transmission and then export the results as a .tif file. For advice on preparing the input data see `vignette("prep-input-data")`.
+This example code demonstrates how to use the package to compute wildfire exposure and remove non-burnable cells in the output.
 
-```{r example1, eval = FALSE}
+We remove non-burnable cells because every cell in the output will be assigned an exposure value regardless of if it can actually recieve embers. Removing these cells (for example: Open Water or Exposed rock) will provide valid summaries and comparisons of the scope of the problem - high exposure of a water cell would inflate the scope of the problem without any rational basis to do so - so we remove those values and only summarize high exposure for the burnable landbase.
+
+
+For advice on preparing your own input data see `vignette("prep-input-data")`.
+
+```{r example1a}
 # Load the fireexposuR library
 library(fireexposuR)
 
 # Load the terra library for reading/writing spatial data
 library(terra)
 
-# read raster of wildland fuels hazardous for long-range ember transmission
-hazard_long <- rast("long_range_hazard.tif")
+# read example hazard data
+hazard_file_path <- "extdata/hazard.tif"
+hazard <- terra::rast(system.file(hazard_file_path, package = "fireexposuR"))
 
-# read raster of wildland fuels hazardous for short-range ember transmission
-hazard_short <- rast("short_range_hazard.tif")
+# read example non-burnable data
+nb_file_path <- "extdata/nb.tif"
+no_burn <- terra::rast(system.file(nb_file_path, package = "fireexposuR"))
 
-# compute long-range ember exposure using the long-range hazard layer
-exposure_long <- fire_exp(hazard_long, t_dist = 500)
-
-# compute short-range ember exposure using the short-range hazard layer
-exposure_short <- fire_exp(hazard_short, t_dist = 100)
-
-# Export the results as a .tif file using terra
-writeRaster(exposure_long, "exposure_long.tif")
-writeRaster(exposure_short, "exposure_short.tif")
 ```
+
+We will compute exposure for all of the cells, and with the non-burnable cells to compare them. 
+
+```{r example1b}
+# compute long-range ember exposure using the long-range hazard raster
+exposure <- fire_exp(hazard, t_dist = 500, no_burn = no_burn)
+
+
+# compute long-range ember exposure without removing non-burnable cells
+exposure_all_cells <- fire_exp(hazard, t_dist = 500)
+```
+
+Here we can compare the difference of removing the non-burnable cells. Its clear that most of the cells that were removed are barren rocky areas in the mountains, but include some man-made features as well like roads.
+
+```{r example1c, fig.width=9, fig.height=8,  fig.show="hold", out.width="50%"}
+# visualize in a map
+fire_exp_map(exposure, title = "Long-range exposure (non-burnable removed)")
+
+fire_exp_map(exposure_all_cells, title = "Long-range exposure (all cells)")
+```
+
+
+Looking at the summary table we can see that removing the non-burnable cells significantly decreases the proportion of cells with no exposure, and increases the proportion of extremely exposed cells by ~10%. 
+
+```{r example1d}
+# visualize in a summary table
+fire_exp_summary(exposure, classify = "landscape")
+
+fire_exp_summary(exposure_all_cells, classify = "landscape")
+
+```
+
+
 
 ### Example: Assess directional exposure
 
@@ -141,45 +173,24 @@ Required input data:
 
 #### Example workflow
 
[…]
