### 2025-09-16 #8479 @mengqi-z
👋 @thomasarsouze - Just checking in to see how the review is going. I believe the authors have addressed all your issues and summarized them in https://github.com/ropensci/fireexposuR/pull/20. Could you please provide a brief status update in this thread? Thanks!

@heyairf - Thanks for addressing the reviewer’s comments! GitHub is always a great place to learn. A couple of suggestions: when you push relevant commits, be sure to tag the issue number in the comments so the commits are automatically linked to the issue. After fixing the technical problem, reply directly on each issue to let the reviewer know what you’ve addressed. Finally, once everything is resolved and the reviewer is satisfied, don’t forget to close the issue.
