### 2025-09-19 ropensci/fireexposuR#19 @heyairf
I have populated this vignette with more info and guidance with #20 . feel free to reopen if you think it needs more
