### 2025-09-19 691abb7
Update fireexposuR.Rmd, typo
