### 2025-09-22 ropensci/fireexposuR#18 @thomasarsouze
Looks good to me. Thanks.
