### 2025-09-29 99fce70
resolution reduced for build file size
