### 2025-10-01 5ca0340
fixes for cran submission
