### 2025-10-17 061148c
#22 JOSS paper edits
