+++ paper/paper.bib
@@ -4,7 +4,7 @@
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
   year = {2024},
-  url = {https://www.R-project.org/},
+  doi = {10.32614/r.manuals}
 }
 
 @article{kim:2024,
