### 2025-10-20 4e1ed0d
Update paper.md figure caption
