### 2025-10-30 #8479 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
